requirements.txt
uvicorn==0.27.0.post1
pydantic==2.6.1
pydantic_core==2.16.2
python-decouple==3.8
python-dotenv==1.0.1
redis==5.0.1
requests==2.31.0
rq==1.15.1
fastapi==0.109.2